+++ app/(auth)/signup/page.tsx
@@ -1,5 +1,16 @@
-import { SignUpView } from "@/modules/auth/ui/sign-up-view";
+import { auth } from "@/lib/auth";
+import { SignUpView } from "@/modules/auth/ui/signup-view";
+import { headers } from "next/headers";
+import { redirect } from "next/navigation";
 
-export default function SignUpPage() {
+export default async function SignUpPage() {
+  const session = await auth.api.getSession({
+    headers: await headers(),
+  });
+
+  if (!!session) {
+    redirect("/");
+  }
+
   return <SignUpView />;
 }
