--- app/(dashboard)/layout.tsx
@@ -1,31 +1,0 @@
-import type React from "react";
-import { SidebarProvider } from "@/components/ui/sidebar";
-import { AppSidebar } from "@/components/app-sidebar";
-import { SidebarInset, SidebarTrigger } from "@/components/ui/sidebar";
-import { ThemeProvider } from "@/components/theme-provider";
-
-export default function Layout({ children }: { children: React.ReactNode }) {
-  return (
-    <ThemeProvider
-      attribute="class"
-      defaultTheme="system"
-      enableSystem
-      disableTransitionOnChange
-    >
-      <SidebarProvider>
-        <AppSidebar />
-        <SidebarInset>
-          <header className="flex h-16 shrink-0 items-center gap-2 border-b px-4">
-            <SidebarTrigger className="-ml-1" />
-            <div className="ml-auto flex items-center space-x-4">
-              <span className="text-sm text-muted-foreground">
-                Welcome back, John!
-              </span>
-            </div>
-          </header>
-          {children}
-        </SidebarInset>
-      </SidebarProvider>
-    </ThemeProvider>
-  );
-}
