--- app/page-auth-temp.tsx
@@ -1,102 +1,0 @@
-// This is a temporary page just to test
-// User creation and
-// User login
-
-"use client";
-
-import { Button } from "@/components/ui/button";
-import { Input } from "@/components/ui/input";
-import { authClient } from "@/lib/auth-client";
-import { useState } from "react";
-
-export default function DashboardPage() {
-  const [name, setName] = useState("");
-  const [email, setEmail] = useState("");
-  const [password, setPassword] = useState("");
-
-  const { data: session } = authClient.useSession();
-
-  const handleSubmit = () => {
-    authClient.signUp.email(
-      {
-        email, // user email address
-        password, // user password -> min 8 characters by default
-        name, // user display name
-      },
-      {
-        onRequest: (ctx) => {
-          //show loading
-          console.log(ctx);
-        },
-        onSuccess: (ctx) => {
-          //redirect to the dashboard or sign in page
-          console.log(ctx);
-        },
-        onError: (ctx) => {
-          // display the error message
-          alert(ctx.error.message);
-        },
-      }
-    );
-  };
-
-  const handleLogin = () => {
-    authClient.signIn.email(
-      {
-        email, // user email address
-        password, // user password -> min 8 characters by default
-      },
-      {
-        onRequest: (ctx) => {
-          //show loading
-          console.log(ctx);
-        },
-        onSuccess: (ctx) => {
-          //redirect to the dashboard or sign in page
-          console.log(ctx);
-        },
-        onError: (ctx) => {
-          // display the error message
-          alert(ctx.error.message);
-        },
-      }
-    );
-  };
-
-  if (session) {
-    return (
-      <div className="p-4 flex gap-2 items-center justify-center">
-        <p>Logged in as {session.user.name}</p>
-        <Button onClick={() => authClient.signOut()}>Sign out</Button>
-      </div>
-    );
-  }
-
-  return (
-    <>
-      <div className="p-4 flex flex-col gap-2">
-        <Input placeholder="Name" onChange={(e) => setName(e.target.value)} />
-        <Input placeholder="Email" onChange={(e) => setEmail(e.target.value)} />
-        <Input
-          placeholder="Password"
-          type="password"
-          onChange={(e) => setPassword(e.target.value)}
-        />
-        <div>
-          <Button onClick={handleSubmit}>Create user</Button>
-        </div>
-      </div>
-      <div className="p-4 flex flex-col gap-2 mt-20">
-        <Input placeholder="Email" onChange={(e) => setEmail(e.target.value)} />
-        <Input
-          placeholder="Password"
-          type="password"
-          onChange={(e) => setPassword(e.target.value)}
-        />
-        <div>
-          <Button onClick={handleLogin}>Log in</Button>
-        </div>
-      </div>
-    </>
-  );
-}
