+++ modules/auth/ui/signup-view.tsx
@@ -1,22 +1,5 @@
 "use client";
 
-import type React from "react";
-
-import { useState } from "react";
-import { Button } from "@/components/ui/button";
-import { Input } from "@/components/ui/input";
-import { Separator } from "@/components/ui/separator";
-import { Alert, AlertTitle } from "@/components/ui/alert";
-import {
-  Eye,
-  EyeOff,
-  Loader2,
-  Mail,
-  Bot,
-  OctagonAlertIcon,
-} from "lucide-react";
-import { useRouter } from "next/navigation";
-import Link from "next/link";
 import {
   Card,
   CardContent,
@@ -24,6 +7,14 @@ import {
   CardHeader,
   CardTitle,
 } from "@/components/ui/card";
+import { Bot, OctagonAlertIcon } from "lucide-react";
+import { useState } from "react";
+import { Button } from "@/components/ui/button";
+import { Input } from "@/components/ui/input";
+import { Separator } from "@/components/ui/separator";
+import { Alert, AlertTitle } from "@/components/ui/alert";
+import { Eye, EyeOff, Loader2, Mail } from "lucide-react";
+import Link from "next/link";
 import { z } from "zod";
 import { useForm } from "react-hook-form";
 import { zodResolver } from "@hookform/resolvers/zod";
@@ -36,39 +27,51 @@ import {
   FormMessage,
 } from "@/components/ui/form";
 import { authClient } from "@/lib/auth-client";
+import { useRouter } from "next/navigation";
 
-const formSchema = z.object({
-  email: z.string().email(),
-  password: z.string().min(1, { message: "Password is required" }),
-});
+const formSchema = z
+  .object({
+    name: z.string().min(1, { message: "Name is required" }),
+    email: z.string().email(),
+    password: z.string().min(1, { message: "Password is required" }),
+    confirmPassword: z.string().min(1, { message: "Password is required" }),
+  })
+  .refine((data) => data.password === data.confirmPassword, {
+    message: "Passwords don't match",
+    path: ["confirmPassword"],
+  });
 
 type FormSchemaType = z.infer<typeof formSchema>;
 
-export function SignInView() {
+export const SignUpView = () => {
+  const router = useRouter();
   const form = useForm<FormSchemaType>({
     resolver: zodResolver(formSchema),
     defaultValues: {
+      name: "",
       email: "",
       password: "",
+      confirmPassword: "",
     },
   });
 
-  const router = useRouter();
   const [isLoading, setIsLoading] = useState(false);
   const [showPassword, setShowPassword] = useState(false);
   const [errors, setErrors] = useState<Record<string, string>>({});
 
   const handleSubmit = (data: FormSchemaType) => {
-    setErrors({});
     setIsLoading(true);
 
-    authClient.signIn.email(
+    authClient.signUp.email(
       {
+        name: data.name,
         email: data.email,
         password: data.password,
+        callbackURL: "/",
       },
       {
         onSuccess: () => {
+          setIsLoading(false);
           router.push("/");
         },
         onError: ({ error }) => {
@@ -79,9 +82,24 @@ export function SignInView() {
     );
   };
 
-  const handleSocialSignIn = (provider: string) => {
-    console.log(`Sign in with ${provider}`);
-    // TODO: Implement social sign in logic
+  const onSocial = (provider: "github" | "google") => {
+    setIsLoading(true);
+
+    authClient.signIn.social(
+      {
+        provider: provider,
+        callbackURL: "/",
+      },
+      {
+        onSuccess: () => {
+          setIsLoading(false);
+        },
+        onError: ({ error }) => {
+          setErrors({ submit: error.message });
+          setIsLoading(false);
+        },
+      }
+    );
   };
 
   return (
@@ -98,29 +116,29 @@ export function SignInView() {
             </div>
           </div>
           <h1 className="text-2xl font-semibold tracking-tight">
-            Welcome back
+            Create your account
           </h1>
           <p className="text-muted-foreground">
-            Sign in to your AI-powered meeting platform
+            Start your AI-powered meeting journey today
           </p>
         </div>
 
-        {/* Sign In Form */}
+        {/* Sign Up Form */}
         <Card>
           <CardHeader className="space-y-1">
-            <CardTitle className="text-xl">Sign in</CardTitle>
+            <CardTitle className="text-xl">Sign up</CardTitle>
             <CardDescription>
-              Enter your email and password to access your account
+              Enter your information to create an account
             </CardDescription>
           </CardHeader>
           <CardContent>
             <div className="space-y-6">
-              {/* Social Sign In */}
+              {/* Social Sign Up */}
               <div className="space-y-3">
                 <Button
                   variant="outline"
                   className="w-full"
-                  onClick={() => handleSocialSignIn("google")}
+                  onClick={() => onSocial("google")}
                   disabled={isLoading}
                 >
                   <Mail className="mr-2 h-4 w-4" />
@@ -129,7 +147,7 @@ export function SignInView() {
                 <Button
                   variant="outline"
                   className="w-full"
-                  onClick={() => handleSocialSignIn("github")}
+                  onClick={() => onSocial("github")}
                   disabled={isLoading}
                 >
                   <svg
@@ -148,15 +166,12 @@ export function SignInView() {
                 </Button>
               </div>
 
-              <div className="relative">
-                <div className="absolute inset-0 flex items-center">
-                  <Separator className="w-full" />
-                </div>
-                <div className="relative flex justify-center text-xs uppercase">
-                  <span className="bg-background px-2 text-muted-foreground">
-                    Or continue with email
-                  </span>
-                </div>
+              <div className="relative flex items-center text-xs uppercase">
+                <Separator className="flex-1" />
+                <span className="px-2 text-muted-foreground flex-none">
+                  Or continue with email
+                </span>
+                <Separator className="flex-1" />
               </div>
 
               {/* Form */}
@@ -165,6 +180,25 @@ export function SignInView() {
                   onSubmit={form.handleSubmit(handleSubmit)}
                   className="space-y-4"
                 >
+                  {/* Name */}
+                  <FormField
+                    control={form.control}
+                    name="name"
+                    render={({ field }) => (
+                      <FormItem>
+                        <FormLabel>Name</FormLabel>
+                        <FormControl>
+                          <Input
+                            placeholder="John Doe"
+                            disabled={isLoading}
+                            {...field}
+                          />
+                        </FormControl>
+                        <FormMessage />
+                      </FormItem>
+                    )}
+                  />
+
                   {/* Email */}
                   <FormField
                     control={form.control}
@@ -191,15 +225,7 @@ export function SignInView() {
                     name="password"
                     render={({ field }) => (
                       <FormItem>
-                        <div className="flex items-center justify-between">
-                          <FormLabel>Password</FormLabel>
-                          <Link
-                            href="/forgot-password"
-                            className="text-sm text-primary hover:underline"
-                          >
-                            Forgot password?
-                          </Link>
-                        </div>
+                        <FormLabel>Password</FormLabel>
                         <FormControl>
                           <div className="relative">
                             <Input
@@ -212,7 +238,43 @@ export function SignInView() {
                               type="button"
                               variant="ghost"
                               size="sm"
-                              className="absolute right-0 top-0 h-full px-3 py-2 hover:bg-transparent"
+                              className="absolute right-0 top-0 h-full px-3 py-2 hover:bg-transparent!"
+                              onClick={() => setShowPassword(!showPassword)}
+                              disabled={isLoading}
+                            >
+                              {showPassword ? (
+                                <EyeOff className="h-4 w-4" />
+                              ) : (
+                                <Eye className="h-4 w-4" />
+                              )}
+                            </Button>
+                          </div>
+                        </FormControl>
+                        <FormMessage />
+                      </FormItem>
+                    )}
+                  />
+
+                  {/* Confirm Password */}
+                  <FormField
+                    control={form.control}
+                    name="confirmPassword"
+                    render={({ field }) => (
+                      <FormItem>
+                        <FormLabel>Confirm Password</FormLabel>
+                        <FormControl>
+                          <div className="relative">
+                            <Input
+                              type={showPassword ? "text" : "password"}
+                              placeholder="********"
+                              disabled={isLoading}
+                              {...field}
+                            />
+                            <Button
+                              type="button"
+                              variant="ghost"
+                              size="sm"
+                              className="absolute right-0 top-0 h-full px-3 py-2 hover:bg-transparent!"
                               onClick={() => setShowPassword(!showPassword)}
                               disabled={isLoading}
                             >
@@ -242,64 +304,52 @@ export function SignInView() {
                     {isLoading ? (
                       <>
                         <Loader2 className="mr-2 h-4 w-4 animate-spin" />
-                        Signing in...
+                        Creating account...
                       </>
                     ) : (
-                      "Sign in"
+                      "Create account"
                     )}
                   </Button>
                 </form>
               </Form>
-
-              {/* Additional Links */}
-              <div className="text-center text-sm text-muted-foreground">
-                <p>
-                  {"Don't have an account? "}
-                  <Link
-                    href="/signup"
-                    className="text-primary hover:underline font-medium"
-                  >
-                    Sign up for free
-                  </Link>
-                </p>
-              </div>
             </div>
           </CardContent>
         </Card>
 
-        {/* Terms */}
-        <div className="text-center text-muted-foreground text-xs *:[a]:hover:underline">
-          By clicking continue, you agree to our{" "}
-          <a href="#" className="text-primary font-medium">
-            Terms of service
-          </a>{" "}
-          and{" "}
-          <a href="#" className="text-primary font-medium">
-            Privacy policy
-          </a>
+        {/* Sign In Link */}
+        <div className="text-center text-sm">
+          <span className="text-muted-foreground">
+            Already have an account?{" "}
+          </span>
+          <Link
+            href="/signin"
+            className="font-medium text-primary hover:underline"
+          >
+            Sign in
+          </Link>
         </div>
 
-        {/* Trust Indicators */}
+        {/* Features Preview */}
         <div className="text-center space-y-4 pt-6">
           <p className="text-sm text-muted-foreground">
-            Trusted by 10,000+ teams worldwide
+            Join thousands of teams using AI SaaS
           </p>
-          <div className="flex justify-center space-x-6 text-xs text-muted-foreground">
-            <div className="flex items-center space-x-1">
-              <div className="w-2 h-2 bg-green-500 rounded-full"></div>
-              <span>99.9% Uptime</span>
+          <div className="grid grid-cols-3 gap-4 text-xs text-muted-foreground">
+            <div className="space-y-1">
+              <div className="font-medium">🤖 AI Agents</div>
+              <div>Smart meeting assistance</div>
             </div>
-            <div className="flex items-center space-x-1">
-              <div className="w-2 h-2 bg-blue-500 rounded-full"></div>
-              <span>SOC 2 Compliant</span>
+            <div className="space-y-1">
+              <div className="font-medium">📞 HD Video</div>
+              <div>Crystal clear calls</div>
             </div>
-            <div className="flex items-center space-x-1">
-              <div className="w-2 h-2 bg-purple-500 rounded-full"></div>
-              <span>GDPR Ready</span>
+            <div className="space-y-1">
+              <div className="font-medium">📝 Auto Notes</div>
+              <div>AI-generated summaries</div>
             </div>
           </div>
         </div>
       </div>
     </div>
   );
-}
+};
