+++ modules/home/ui/views/home-view.tsx
@@ -1,0 +1,29 @@
+"use client";
+
+import { Button } from "@/components/ui/button";
+import { authClient } from "@/lib/auth-client";
+import { useRouter } from "next/navigation";
+
+export const HomeView = () => {
+  const router = useRouter();
+  const { data: session } = authClient.useSession();
+
+  if (!session) {
+    return <p>Loading...</p>;
+  }
+
+  return (
+    <div>
+      <p>Logged in as {session.user.name}</p>
+      <Button
+        onClick={() =>
+          authClient.signOut({
+            fetchOptions: { onSuccess: () => router.push("/signin") },
+          })
+        }
+      >
+        Sign out
+      </Button>
+    </div>
+  );
+};
